WIP: Style ImportPhrase
