+++ src/components/uielements/backLink/BackLink.stories.tsx
@@ -7,7 +7,7 @@ import BackLink from './BackLink'
 storiesOf('Components/BackLink', module).add('default', () => {
   return (
     <div style={{ padding: '15px' }}>
-      <BackLink route="/" />
+      <BackLink />
     </div>
   )
 })
