+++ src/components/uielements/backLink/BackLink.tsx
@@ -1,25 +1,25 @@
 import React from 'react'
 
 import { LeftOutlined } from '@ant-design/icons'
-import { Link } from 'react-router-dom'
+import { useHistory } from 'react-router-dom'
 
 import { BackLinkWrapper } from './BackLink.style'
 
 type Props = {
-  route: string
   title?: string
 }
 
-const BackLink: React.FC<Props> = ({ route, title = 'Back', ...otherProps }): JSX.Element => {
+const BackLink: React.FC<Props> = ({ title = 'Back', ...otherProps }): JSX.Element => {
+  const history = useHistory()
+
+  const clickHandler = () => {
+    history.goBack()
+  }
   return (
-    <div {...otherProps}>
-      <Link to={route}>
-        <BackLinkWrapper>
-          <LeftOutlined />
-          <span>{title}</span>
-        </BackLinkWrapper>
-      </Link>
-    </div>
+    <BackLinkWrapper {...otherProps} onClick={clickHandler}>
+      <LeftOutlined />
+      <span>{title}</span>
+    </BackLinkWrapper>
   )
 }
