+++ src/views/wallet/WalletView.tsx
@@ -11,7 +11,7 @@ import { hasImportedKeystore, isLocked } from '../../services/wallet/util'
 import View from '../View'
 import FundsReceiveScreen from './FundsReceiveScreen'
 import FundsSendScreen from './FundsSendScreen'
-import ImportsView from './ImportsView'
+import ImportsView from './importsView/ImportsView'
 import UnlockView from './UnlockView'
 import UserAssetDetailsScreen from './UserAssetDetailsScreen'
 import UserAssetsScreen from './UserAssetsScreen'
