+++ src/views/wallet/importsView/ImportsView.style.tsx
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+export const ImportsViewWrapper = styled.div`
+  .tabs {
+  }
+`
