+++ src/views/wallet/importsView/index.tsx
@@ -1,0 +1,3 @@
+import ImportsView from './ImportsView'
+
+export default ImportsView
